pyproject.toml
[tool.poetry]
name = "dispike"
version = "0.2.2-alpha.0"
description = ""
authors = ["Mustafa Mohamed <ms7mohamed@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
loguru = "^0.5.3"
fastapi = "^0.63.0"
pydantic = "^1.7.3"
typing-extensions = "^3.7.4"
PyNaCl = "^1.4.0"
httpx = "^0.16.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"
mypy = "^0.790"
uvicorn = "^0.13.2"
requests = "^2.25.1"
pytest-asyncio = "^0.14.0"
pylint = "^2.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
